% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cornsoybean}
\alias{cornsoybean}
\title{Corn and soy beans survey and satellite data in 12 counties in Iowa}
\format{
A data frame with 37 observations on the following 5 variables.
\describe{
   \item{\code{County}:}{ numeric county code.}
   \item{\code{CornHec}:}{ reported hectares of corn from the survey.}
   \item{\code{SoyBeansHec}:}{ reported hectares of soy beans from the survey.}
   \item{\code{CornPix}:}{ number of pixels of corn in sample segment within county, from satellite data. }
   \item{\code{SoyBeansPix}:}{ number of pixels of soy beans in sample segment within county, from satellite data.}
}
}
\source{
Battesse, G.E., Harter, R.M. and Fuller, W.A. (1988). An Error-Components Model for Prediction of County Crop Areas Using Survey and Satellite Data. Journal of the American Statistical Association 83, 28-36.
}
\usage{
cornsoybean
}
\description{
Survey and satellite data for corn and soy beans in 12 Iowa counties, obtained from the 1978 June Enumerative Survey of the U.S. Department of Agriculture and from land observatory satellites (LANDSAT) during the 1978 growing season.
}
\details{
cornsoybean
}
\keyword{datasets}
