% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cornsoybeanmeans}
\alias{cornsoybeanmeans}
\title{cornsoybeanmeans}
\format{
An object of class \code{data.frame} with 12 rows and 6 columns.
}
\usage{
cornsoybeanmeans
}
\description{
cornsoybeanmeans
}
\keyword{datasets}
