% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy_area}
\alias{dummy_area}
\title{dummy_area}
\format{
A data frame with 30 observations on the following 4 variables.
\describe{
   \item{\code{domain}:}{domain code}
   \item{\code{x1}:}{x1}
   \item{\code{x2}:}{x2}
   \item{\code{parameter}:}{ true value of y }
}
}
\usage{
dummy_area
}
\description{
dummy data
}
\details{
dummy_area
}
\keyword{datasets}
