% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy_unit}
\alias{dummy_unit}
\title{dummy_unit}
\format{
A data frame with 1000 observations on the following 4 variables.
\describe{
   \item{\code{domain}:}{domain code}
   \item{\code{y_di}:}{direct estimate of y}
   \item{\code{x1}:}{x1}
   \item{\code{x2}:}{x2}
}
}
\usage{
dummy_unit
}
\description{
dummy data
}
\details{
dummy_unit
}
\keyword{datasets}
