% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cornsoybean}
\alias{cornsoybean}
\title{cornsoybean}
\format{
An object of class \code{data.frame} with 37 rows and 5 columns.
}
\usage{
cornsoybean
}
\description{
cornsoybean
}
\keyword{datasets}
