% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cornsoybeanmeans}
\alias{cornsoybeanmeans}
\title{Corn and soy beans mean number of pixels per segment for 12 counties in Iowa.}
\format{
A data frame with 12 observations on the following 6 variables.
\describe{
     \item{\code{CountyIndex}:}{ numeric county code.}
     \item{\code{CountyName}:}{ name of the county.}
     \item{\code{SampSegments}:}{ number of sample segments in the county (sample size).}
     \item{\code{PopnSegments}:}{ number of population segments in the county (population size).}
     \item{\code{MeanCornPixPerSeg}:}{ mean number of corn pixels per segment in the county.}
     \item{\code{MeanSoyBeansPixPerSeg}:}{ mean number of soy beans pixels per segment in the county.}
 }
}
\source{
Battesse, G.E., Harter, R.M. and Fuller, W.A. (1988). An Error-Components Model for Prediction of County Crop Areas Using Survey and Satellite Data. Journal of the American Statistical Association 83, 28-36.
}
\usage{
cornsoybeanmeans
}
\description{
County means of number of pixels per segment of corn and soy beans, from satellite data, for 12 counties in Iowa. Population size, sample size and means of auxiliary variables in data set \code{\link{cornsoybean}}.
}
\details{
cornsoybeanmeans
}
\keyword{datasets}
